requirements.txt
anyio==4.13.0
asyncpg==0.31.0
click==8.3.2
et_xmlfile==2.0.0
greenlet==3.4.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing_extensions==4.15.0
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
